move swagger docs to correct /v1 section
