--- app/swagger/swagger/schemas/gibct/version_public_exports.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Swagger
-  module Schemas
-    module Gibct
-      class VersionPublicExports
-        include Swagger::Blocks
-      end
-    end
-  end
-end
