+++ app/swagger/swagger/v1/schemas/gibct/version_public_exports.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Swagger
+  module V1
+    module Schemas
+      module Gibct
+        class VersionPublicExports
+          include Swagger::Blocks
+        end
+      end
+    end
+  end
+end
